correction delete et modify sauce

+++ controllers/sauces.js
@@ -34,27 +34,43 @@ exports.getOneSauce = (req, res, next) => {
   );
 };
 
-exports.modifySauce = (req, res, next) => {
-  const hotSauce = req.file ?
-  {
-    ...JSON.parse(req.body.sauce),
-    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`} : { ...req.body }
+exports.modifySauce = (req, res) => {
+  let hotSauce = {};
+  req.file ? (
+      Sauce.findOne({_id: req.params.id})
+          .then(sauce => {
+              if(sauce.userId !== req.auth.userId){
+                  res.status(403).json({message: `Non autorisé !`})
+              } else {
+                  const filename = sauce.imageUrl.split("/").at(-1);
+                  fs.unlinkSync(`images/${filename}`)
+              }
+          }),
+      hotSauce = {
+          ...JSON.parse(req.body.sauce),
+          imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
+          }
+      ) : sauceObject = {...req.body};
+  Sauce.updateOne({_id: req.params.id},{...sauceObject, _id: req.params.id})
+      .then(() => res.status(200).json({message: 'Sauce modifiée !'}))
+      .catch((error) => res.status(400).json({error}))
+};
 
-  Sauce.updateOne({ _id : req.params.id}, {...hotSauce, _id: req.params.id})
-  .then(res.status(200).json({ message : "Sauce modifiée"}))
-  .catch(error => res.status(400).json({ error }))
-}
-
-exports.deleteSauce = (req, res, next) => {
+exports.deleteSauce = (req, res) => {
+  console.log(req.auth)
   Sauce.findOne({ _id: req.params.id })
-    .then((sauce) => {
+  .then (sauce => {
+    console.log('sauce:', sauce);
+    if (sauce.userId !== req.auth.userId) {
+        res.status(403).json({message: `Non autorisé`});
+      } else {
       const filename = sauce.imageUrl.split('/images/')[1];
       fs.unlink(`images/${filename}`, () => {
         Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: 'Sauce supprimée !' }))
+          .then(() => res.status(200).json({ message: `Sauce supprimée !` }))
           .catch((error) => res.status(400).json({ error }));
       });
-    })
+    }})
     .catch((error) => res.status(500).json({ error }));
 };
 
@@ -86,16 +102,15 @@ exports.likeDislikeSauce = (req, res, next) => {
             
       break;
 
-    case 0 :
+      case 0:
         Sauce.findOne({ _id: sauceId })
-           .then((sauce) => {
-            if (sauce.usersLiked.includes(userId)) { 
-              Sauce.updateOne({ _id: sauceId }, { $pull: { usersLiked: userId }, $inc: { likes: -1 }})
+          .then((sauce) => {
+            if (sauce.usersLiked.includes(userId)) {
+              Sauce.updateOne({ _id: sauceId }, { $pull: { usersLiked: userId }, $inc: { likes: -1 } })
                 .then(() => res.status(200).json({ message: `Neutre` }))
                 .catch((error) => res.status(400).json({ error }))
-            }
-            if (sauce.usersDisliked.includes(userId)) { 
-              Sauce.updateOne({ _id: sauceId }, { $pull: { usersDisliked: userId }, $inc: { dislikes: -1 }})
+            } else if (sauce.usersDisliked.includes(userId)) {
+              Sauce.updateOne({ _id: sauceId }, { $pull: { usersDisliked: userId }, $inc: { dislikes: -1 } })
                 .then(() => res.status(200).json({ message: `Neutre` }))
                 .catch((error) => res.status(400).json({ error }))
             }
